Fix workflow notifier by restructuring
